+++ main.py
@@ -1,9 +1,9 @@
 import abcExample
-import AlphabetV2
+import alphabet
 
 if __name__ == "__main__":
     print("hello world")
     start = input("Enter a start string: ")  # aaa
     finish = input("Enter where to finish: ")  # aax
     abcExample.abc(start, finish)
-    AlphabetV2.alphabetV2(start, finish)
+    alphabet.alphabetV2(start, finish)
